improv: clean up code
