+++ gui/src/redux/thunks/streamResponse.test.ts
@@ -2590,2112 +2590,4 @@ describe("streamResponseThunk", () => {
       },
     });
   });
-
-  it("should handle tool call rejection error", async () => {
-    const { mockStore, mockIdeMessenger } = setupTest();
-
-    // Create store with tool settings that will reject the tool call
-    const mockStoreWithToolReject = createMockStore({
-      session: {
-        history: [
-          {
-            message: {
-              id: "1",
-              role: "user",
-              content: "Please edit this file",
-            },
-            contextItems: [],
-          },
-        ],
-        hasReasoningEnabled: false,
-        isStreaming: true,
-        id: "session-123",
-        mode: "chat",
-        streamAborter: new AbortController(),
-        contextPercentage: 0,
-        isPruned: false,
-        isInEdit: false,
-        title: "",
-        lastSessionId: undefined,
-        isSessionMetadataLoading: false,
-        allSessionMetadata: [],
-        symbols: {},
-        codeBlockApplyStates: {
-          states: [],
-          curIndex: 0,
-        },
-        newestToolbarPreviewForInput: {},
-        compactionLoading: {},
-        inlineErrorMessage: undefined,
-      },
-      config: {
-        config: {
-          tools: [],
-          rules: [],
-          tabAutocompleteModel: undefined,
-          selectedModelByRole: {
-            chat: mockClaudeModel,
-            apply: null,
-            edit: null,
-            summarize: null,
-            autocomplete: null,
-            rerank: null,
-            embed: null,
-          },
-          experimental: {
-            onlyUseSystemMessageTools: false,
-          },
-        } as any,
-        lastSelectedModelByRole: {
-          chat: mockClaudeModel.title,
-        },
-      } as any,
-      ui: {
-        toolSettings: {
-          edit_existing_file: "disabled", // Tool is disabled/rejected
-        },
-        ruleSettings: {},
-        showDialog: false,
-        dialogMessage: undefined,
-      } as any,
-    });
-
-    const mockIdeMessengerReject = mockStoreWithToolReject.mockIdeMessenger;
-
-    // Setup successful compilation
-    mockIdeMessengerReject.request.mockResolvedValue({
-      status: "success",
-      content: {
-        compiledChatMessages: [
-          { role: "user", content: "Please edit this file" },
-        ],
-        didPrune: false,
-        contextPercentage: 0.7,
-      },
-    });
-
-    // Setup streaming generator with tool call that will be rejected
-    async function* mockStreamGeneratorWithRejectedTool() {
-      yield [{ role: "assistant", content: "I'll edit the file for you." }];
-      yield [
-        {
-          role: "assistant",
-          content: "",
-          toolCalls: [
-            {
-              id: "tool-reject-1",
-              type: "function",
-              function: {
-                name: "edit_existing_file",
-                arguments: JSON.stringify({
-                  filepath: "test.js",
-                  changes: "const x = 1;",
-                }),
-              },
-            },
-          ],
-        },
-      ];
-      return {
-        prompt: "Please edit this file",
-        completion: "I'll edit the file for you.",
-        modelProvider: "anthropic",
-      };
-    }
-
-    mockIdeMessengerReject.llmStreamChat.mockReturnValue(
-      mockStreamGeneratorWithRejectedTool(),
-    );
-
-    // Execute thunk
-    const result = await mockStoreWithToolReject.dispatch(
-      streamResponseThunk({
-        editorState: mockEditorState,
-        modifiers: mockModifiers,
-      }) as any,
-    );
-
-    // Verify thunk completed successfully (tool rejection is handled gracefully)
-    expect(result.type).toBe("chat/streamResponse/fulfilled");
-
-    // Verify exact action sequence includes tool generation but rejection handling
-    const dispatchedActions = (mockStoreWithToolReject as any).getActions();
-    expect(dispatchedActions).toEqual([
-      {
-        type: "chat/streamResponse/pending",
-        meta: {
-          arg: {
-            editorState: mockEditorState,
-            modifiers: mockModifiers,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/pending",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/submitEditorAndInitAtIndex",
-        payload: {
-          editorState: mockEditorState,
-          index: 1,
-        },
-      },
-      {
-        type: "session/resetNextCodeBlockToApplyIndex",
-        payload: undefined,
-      },
-      {
-        type: "symbols/updateFromContextItems/pending",
-        meta: {
-          arg: [],
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/updateHistoryItemAtIndex",
-        payload: {
-          index: 1,
-          updates: {
-            contextItems: [],
-            message: {
-              content: "Hello, please help me with this code",
-              id: "mock-uuid-123",
-              role: "user",
-            },
-          },
-        },
-      },
-      {
-        type: "chat/streamNormalInput/pending",
-        meta: {
-          arg: {
-            legacySlashCommandData: undefined,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setAppliedRulesAtIndex",
-        payload: {
-          appliedRules: [],
-          index: 1,
-        },
-      },
-      {
-        type: "session/setActive",
-        payload: undefined,
-      },
-      {
-        type: "session/setInlineErrorMessage",
-        payload: undefined,
-      },
-      {
-        type: "session/setIsPruned",
-        payload: false,
-      },
-      {
-        type: "session/setContextPercentage",
-        payload: 0.7,
-      },
-      {
-        type: "symbols/updateFromContextItems/fulfilled",
-        meta: {
-          arg: [],
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content: "I'll edit the file for you.",
-          },
-        ],
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content: "",
-            toolCalls: [
-              {
-                id: "tool-reject-1",
-                type: "function",
-                function: {
-                  name: "edit_existing_file",
-                  arguments: '{"filepath":"test.js","changes":"const x = 1;"}',
-                },
-              },
-            ],
-          },
-        ],
-      },
-      {
-        type: "session/addPromptCompletionPair",
-        payload: [
-          {
-            completion: "I'll edit the file for you.",
-            modelProvider: "anthropic",
-            prompt: "Please edit this file",
-          },
-        ],
-      },
-      {
-        type: "session/setInactive",
-        payload: undefined,
-      },
-      {
-        type: "session/setToolGenerated",
-        payload: {
-          toolCallId: "tool-reject-1",
-          tools: [],
-        },
-      },
-      {
-        type: "chat/streamNormalInput/fulfilled",
-        meta: {
-          arg: {
-            legacySlashCommandData: undefined,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/pending",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/update/pending",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/updateSessionMetadata",
-        payload: {
-          sessionId: "session-123",
-          title: "New Session",
-        },
-      },
-      {
-        type: "session/refreshMetadata/pending",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setIsSessionMetadataLoading",
-        payload: false,
-      },
-      {
-        type: "session/setAllSessionMetadata",
-        payload: {
-          compiledChatMessages: [
-            {
-              content: "Please edit this file",
-              role: "user",
-            },
-          ],
-          contextPercentage: 0.7,
-          didPrune: false,
-        },
-      },
-      {
-        type: "session/refreshMetadata/fulfilled",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: {
-          compiledChatMessages: [
-            {
-              content: "Please edit this file",
-              role: "user",
-            },
-          ],
-          contextPercentage: 0.7,
-          didPrune: false,
-        },
-      },
-      {
-        type: "session/update/fulfilled",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/fulfilled",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/fulfilled",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamResponse/fulfilled",
-        meta: {
-          arg: {
-            editorState: mockEditorState,
-            modifiers: mockModifiers,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-    ]);
-
-    // Verify IDE messenger calls - compilation should succeed, streaming should happen
-    expect(mockIdeMessengerReject.request).toHaveBeenCalledWith(
-      "llm/compileChat",
-      {
-        messages: [
-          {
-            role: "system",
-            content: "You are a helpful assistant.",
-          },
-          {
-            role: "user",
-            content: [
-              {
-                type: "text",
-                text: "Please edit this file",
-              },
-            ],
-          },
-          {
-            role: "user",
-            content: [
-              {
-                type: "text",
-                text: "Hello, please help me with this code",
-              },
-            ],
-          },
-        ],
-        options: {},
-      },
-    );
-
-    expect(mockIdeMessengerReject.llmStreamChat).toHaveBeenCalledWith(
-      {
-        completionOptions: {},
-        legacySlashCommandData: undefined,
-        messageOptions: { precompiled: true },
-        messages: [
-          {
-            role: "user",
-            content: "Please edit this file",
-          },
-        ],
-        title: "Claude 3.5 Sonnet",
-      },
-      expect.any(AbortSignal),
-    );
-
-    // Should NOT call tools/call since tool is disabled/rejected
-    expect(mockIdeMessengerReject.request).not.toHaveBeenCalledWith(
-      "tools/call",
-      expect.anything(),
-    );
-
-    // Verify session save was called
-    expect(mockIdeMessengerReject.request).toHaveBeenCalledWith(
-      "history/save",
-      expect.anything(),
-    );
-
-    // Verify final state contains the tool call but it's not executed
-    const finalState = mockStoreWithToolReject.getState();
-    expect(finalState).toEqual({
-      session: {
-        history: [
-          {
-            contextItems: [],
-            message: {
-              id: "1",
-              role: "user",
-              content: "Please edit this file",
-            },
-          },
-          {
-            appliedRules: [],
-            contextItems: [],
-            editorState: mockEditorState,
-            message: {
-              content: "Hello, please help me with this code",
-              id: expect.any(String),
-              role: "user",
-            },
-          },
-          {
-            contextItems: [],
-            isGatheringContext: false,
-            message: {
-              content: "I'll edit the file for you.",
-              id: "mock-uuid-123",
-              role: "assistant",
-              toolCalls: [
-                {
-                  id: "tool-reject-1",
-                  type: "function",
-                  function: {
-                    name: "edit_existing_file",
-                    arguments: JSON.stringify({
-                      filepath: "test.js",
-                      changes: "const x = 1;",
-                    }),
-                  },
-                },
-              ],
-            },
-            promptLogs: [
-              {
-                completion: "I'll edit the file for you.",
-                modelProvider: "anthropic",
-                prompt: "Please edit this file",
-              },
-            ],
-            toolCallStates: [
-              {
-                toolCallId: "tool-reject-1",
-                toolCall: {
-                  id: "tool-reject-1",
-                  type: "function",
-                  function: {
-                    name: "edit_existing_file",
-                    arguments: JSON.stringify({
-                      filepath: "test.js",
-                      changes: "const x = 1;",
-                    }),
-                  },
-                },
-                parsedArgs: {
-                  filepath: "test.js",
-                  changes: "const x = 1;",
-                },
-                status: "generated", // Tool call exists but isn't executed due to settings
-              },
-            ],
-          },
-        ],
-        hasReasoningEnabled: false,
-        isStreaming: false,
-        id: "session-123",
-        mode: "chat",
-        streamAborter: expect.any(AbortController),
-        contextPercentage: 0.7,
-        isPruned: false,
-        isInEdit: false,
-        title: "New Session",
-        lastSessionId: undefined,
-        isSessionMetadataLoading: false,
-        allSessionMetadata: {
-          compiledChatMessages: [
-            {
-              content: "Please edit this file",
-              role: "user",
-            },
-          ],
-          contextPercentage: 0.7,
-          didPrune: false,
-        },
-        symbols: {},
-        codeBlockApplyStates: {
-          states: [],
-          curIndex: 0,
-        },
-        newestToolbarPreviewForInput: {},
-        compactionLoading: {},
-        inlineErrorMessage: undefined,
-      },
-      config: {
-        config: {
-          tools: [],
-          rules: [],
-          tabAutocompleteModel: undefined,
-          selectedModelByRole: {
-            chat: mockClaudeModel,
-            apply: null,
-            edit: null,
-            summarize: null,
-            autocomplete: null,
-            rerank: null,
-            embed: null,
-          },
-          experimental: {
-            onlyUseSystemMessageTools: false,
-          },
-        },
-        lastSelectedModelByRole: {
-          chat: mockClaudeModel.title,
-        },
-        loading: false,
-        configError: undefined,
-      },
-      ui: {
-        toolSettings: {
-          edit_existing_file: "disabled",
-        },
-        ruleSettings: {},
-        showDialog: false,
-        dialogMessage: undefined,
-      },
-      editModeState: {
-        isInEdit: false,
-        returnToMode: "chat",
-      },
-      indexing: {
-        indexingState: "disabled",
-      },
-      tabs: {
-        tabsItems: [],
-      },
-      profiles: {
-        profiles: [],
-      },
-    });
-  });
-
-  it("should handle tool call requiring manual approval", async () => {
-    const { mockStore, mockIdeMessenger } = setupTest();
-
-    // Create store with tool settings that require manual approval (ask first)
-    const mockStoreWithManualApproval = createMockStore({
-      session: {
-        history: [
-          {
-            message: {
-              id: "1",
-              role: "user",
-              content: "Please search the codebase",
-            },
-            contextItems: [],
-          },
-        ],
-        hasReasoningEnabled: false,
-        isStreaming: true,
-        id: "session-123",
-        mode: "chat",
-        streamAborter: new AbortController(),
-        contextPercentage: 0,
-        isPruned: false,
-        isInEdit: false,
-        title: "",
-        lastSessionId: undefined,
-        isSessionMetadataLoading: false,
-        allSessionMetadata: [],
-        symbols: {},
-        codeBlockApplyStates: {
-          states: [],
-          curIndex: 0,
-        },
-        newestToolbarPreviewForInput: {},
-        compactionLoading: {},
-        inlineErrorMessage: undefined,
-      },
-      config: {
-        config: {
-          tools: [],
-          rules: [],
-          tabAutocompleteModel: undefined,
-          selectedModelByRole: {
-            chat: mockClaudeModel,
-            apply: null,
-            edit: null,
-            summarize: null,
-            autocomplete: null,
-            rerank: null,
-            embed: null,
-          },
-          experimental: {
-            onlyUseSystemMessageTools: false,
-          },
-        } as any,
-        lastSelectedModelByRole: {
-          chat: mockClaudeModel.title,
-        },
-      } as any,
-      ui: {
-        toolSettings: {
-          search_codebase: "askFirst", // Requires manual approval
-        },
-        ruleSettings: {},
-        showDialog: false,
-        dialogMessage: undefined,
-      } as any,
-    });
-
-    const mockIdeMessengerManual = mockStoreWithManualApproval.mockIdeMessenger;
-
-    // Setup successful compilation
-    mockIdeMessengerManual.request.mockResolvedValue({
-      status: "success",
-      content: {
-        compiledChatMessages: [
-          { role: "user", content: "Please search the codebase" },
-        ],
-        didPrune: false,
-        contextPercentage: 0.9,
-      },
-    });
-
-    // Setup streaming generator with tool call requiring approval
-    async function* mockStreamGeneratorWithApprovalTool() {
-      yield [
-        { role: "assistant", content: "I'll search the codebase for you." },
-      ];
-      yield [
-        {
-          role: "assistant",
-          content: "",
-          toolCalls: [
-            {
-              id: "tool-approval-1",
-              type: "function",
-              function: {
-                name: "search_codebase",
-                arguments: JSON.stringify({ query: "test function" }),
-              },
-            },
-          ],
-        },
-      ];
-      return {
-        prompt: "Please search the codebase",
-        completion: "I'll search the codebase for you.",
-        modelProvider: "anthropic",
-      };
-    }
-
-    mockIdeMessengerManual.llmStreamChat.mockReturnValue(
-      mockStreamGeneratorWithApprovalTool(),
-    );
-
-    // Execute thunk
-    const result = await mockStoreWithManualApproval.dispatch(
-      streamResponseThunk({
-        editorState: mockEditorState,
-        modifiers: mockModifiers,
-      }) as any,
-    );
-
-    // Verify thunk completed successfully (tool waits for approval)
-    expect(result.type).toBe("chat/streamResponse/fulfilled");
-
-    // Verify exact action sequence includes tool generation but NO execution
-    const dispatchedActions = (mockStoreWithManualApproval as any).getActions();
-    expect(dispatchedActions).toEqual([
-      {
-        type: "chat/streamResponse/pending",
-        meta: {
-          arg: {
-            editorState: mockEditorState,
-            modifiers: mockModifiers,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/pending",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/submitEditorAndInitAtIndex",
-        payload: {
-          editorState: mockEditorState,
-          index: 1,
-        },
-      },
-      {
-        type: "session/resetNextCodeBlockToApplyIndex",
-        payload: undefined,
-      },
-      {
-        type: "symbols/updateFromContextItems/pending",
-        meta: {
-          arg: [],
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/updateHistoryItemAtIndex",
-        payload: {
-          index: 1,
-          updates: {
-            contextItems: [],
-            message: {
-              content: "Hello, please help me with this code",
-              id: "mock-uuid-123",
-              role: "user",
-            },
-          },
-        },
-      },
-      {
-        type: "chat/streamNormalInput/pending",
-        meta: {
-          arg: {
-            legacySlashCommandData: undefined,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setAppliedRulesAtIndex",
-        payload: {
-          appliedRules: [],
-          index: 1,
-        },
-      },
-      {
-        type: "session/setActive",
-        payload: undefined,
-      },
-      {
-        type: "session/setInlineErrorMessage",
-        payload: undefined,
-      },
-      {
-        type: "session/setIsPruned",
-        payload: false,
-      },
-      {
-        type: "session/setContextPercentage",
-        payload: 0.9,
-      },
-      {
-        type: "symbols/updateFromContextItems/fulfilled",
-        meta: {
-          arg: [],
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content: "I'll search the codebase for you.",
-          },
-        ],
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content: "",
-            toolCalls: [
-              {
-                id: "tool-approval-1",
-                type: "function",
-                function: {
-                  name: "search_codebase",
-                  arguments: JSON.stringify({ query: "test function" }),
-                },
-              },
-            ],
-          },
-        ],
-      },
-      {
-        type: "session/addPromptCompletionPair",
-        payload: [
-          {
-            completion: "I'll search the codebase for you.",
-            modelProvider: "anthropic",
-            prompt: "Please search the codebase",
-          },
-        ],
-      },
-      {
-        type: "session/setInactive",
-        payload: undefined,
-      },
-      {
-        type: "session/setToolGenerated",
-        payload: {
-          toolCallId: "tool-approval-1",
-          tools: [],
-        },
-      },
-      {
-        type: "chat/streamNormalInput/fulfilled",
-        meta: {
-          arg: {
-            legacySlashCommandData: undefined,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/pending",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/update/pending",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/updateSessionMetadata",
-        payload: {
-          sessionId: "session-123",
-          title: "New Session",
-        },
-      },
-      {
-        type: "session/refreshMetadata/pending",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setIsSessionMetadataLoading",
-        payload: false,
-      },
-      {
-        type: "session/setAllSessionMetadata",
-        payload: {
-          compiledChatMessages: [
-            {
-              content: "Please search the codebase",
-              role: "user",
-            },
-          ],
-          contextPercentage: 0.9,
-          didPrune: false,
-        },
-      },
-      {
-        type: "session/refreshMetadata/fulfilled",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: {
-          compiledChatMessages: [
-            {
-              content: "Please search the codebase",
-              role: "user",
-            },
-          ],
-          contextPercentage: 0.9,
-          didPrune: false,
-        },
-      },
-      {
-        type: "session/update/fulfilled",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/fulfilled",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/fulfilled",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamResponse/fulfilled",
-        meta: {
-          arg: {
-            editorState: mockEditorState,
-            modifiers: mockModifiers,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-    ]);
-
-    // Verify IDE messenger calls - compilation should happen, streaming should happen
-    expect(mockIdeMessengerManual.request).toHaveBeenCalledWith(
-      "llm/compileChat",
-      {
-        messages: [
-          {
-            role: "system",
-            content: "You are a helpful assistant.",
-          },
-          {
-            role: "user",
-            content: [
-              {
-                type: "text",
-                text: "Please search the codebase",
-              },
-            ],
-          },
-          {
-            role: "user",
-            content: [
-              {
-                type: "text",
-                text: "Hello, please help me with this code",
-              },
-            ],
-          },
-        ],
-        options: {},
-      },
-    );
-
-    expect(mockIdeMessengerManual.llmStreamChat).toHaveBeenCalledWith(
-      {
-        completionOptions: {},
-        legacySlashCommandData: undefined,
-        messageOptions: { precompiled: true },
-        messages: [
-          {
-            role: "user",
-            content: "Please search the codebase",
-          },
-        ],
-        title: "Claude 3.5 Sonnet",
-      },
-      expect.any(AbortSignal),
-    );
-
-    // Should NOT call tools/call since tool requires approval
-    expect(mockIdeMessengerManual.request).not.toHaveBeenCalledWith(
-      "tools/call",
-      expect.anything(),
-    );
-
-    // Verify session save was called
-    expect(mockIdeMessengerManual.request).toHaveBeenCalledWith(
-      "history/save",
-      expect.anything(),
-    );
-
-    // Verify final state contains the tool call in "generated" state (waiting for approval)
-    const finalState = mockStoreWithManualApproval.getState();
-    expect(finalState).toEqual({
-      session: {
-        history: [
-          {
-            contextItems: [],
-            message: {
-              id: "1",
-              role: "user",
-              content: "Please search the codebase",
-            },
-          },
-          {
-            appliedRules: [],
-            contextItems: [],
-            editorState: mockEditorState,
-            message: {
-              content: "Hello, please help me with this code",
-              id: expect.any(String),
-              role: "user",
-            },
-          },
-          {
-            contextItems: [],
-            isGatheringContext: false,
-            message: {
-              content: "I'll search the codebase for you.",
-              id: "mock-uuid-123",
-              role: "assistant",
-              toolCalls: [
-                {
-                  id: "tool-approval-1",
-                  type: "function",
-                  function: {
-                    name: "search_codebase",
-                    arguments: JSON.stringify({ query: "test function" }),
-                  },
-                },
-              ],
-            },
-            promptLogs: [
-              {
-                completion: "I'll search the codebase for you.",
-                modelProvider: "anthropic",
-                prompt: "Please search the codebase",
-              },
-            ],
-            toolCallStates: [
-              {
-                toolCallId: "tool-approval-1",
-                toolCall: {
-                  id: "tool-approval-1",
-                  type: "function",
-                  function: {
-                    name: "search_codebase",
-                    arguments: JSON.stringify({ query: "test function" }),
-                  },
-                },
-                parsedArgs: { query: "test function" },
-                status: "generated", // Waiting for user approval
-              },
-            ],
-          },
-        ],
-        hasReasoningEnabled: false,
-        isStreaming: false,
-        id: "session-123",
-        mode: "chat",
-        streamAborter: expect.any(AbortController),
-        contextPercentage: 0.9,
-        isPruned: false,
-        isInEdit: false,
-        title: "New Session",
-        lastSessionId: undefined,
-        isSessionMetadataLoading: false,
-        allSessionMetadata: {
-          compiledChatMessages: [
-            {
-              content: "Please search the codebase",
-              role: "user",
-            },
-          ],
-          contextPercentage: 0.9,
-          didPrune: false,
-        },
-        symbols: {},
-        codeBlockApplyStates: {
-          states: [],
-          curIndex: 0,
-        },
-        newestToolbarPreviewForInput: {},
-        compactionLoading: {},
-        inlineErrorMessage: undefined,
-      },
-      config: {
-        config: {
-          tools: [],
-          rules: [],
-          tabAutocompleteModel: undefined,
-          selectedModelByRole: {
-            chat: mockClaudeModel,
-            apply: null,
-            edit: null,
-            summarize: null,
-            autocomplete: null,
-            rerank: null,
-            embed: null,
-          },
-          experimental: {
-            onlyUseSystemMessageTools: false,
-          },
-        },
-        lastSelectedModelByRole: {
-          chat: mockClaudeModel.title,
-        },
-        loading: false,
-        configError: undefined,
-      },
-      ui: {
-        toolSettings: {
-          search_codebase: "askFirst",
-        },
-        ruleSettings: {},
-        showDialog: false,
-        dialogMessage: undefined,
-      },
-      editModeState: {
-        isInEdit: false,
-        returnToMode: "chat",
-      },
-      indexing: {
-        indexingState: "disabled",
-      },
-      tabs: {
-        tabsItems: [],
-      },
-      profiles: {
-        profiles: [],
-      },
-    });
-  });
-
-  it("should handle complete user approval and tool execution flow", async () => {
-    const { mockStore, mockIdeMessenger } = setupTest();
-
-    // Create store with tool settings that require manual approval
-    const mockStoreWithApproval = createMockStore({
-      session: {
-        history: [
-          {
-            message: {
-              id: "1",
-              role: "user",
-              content: "Please search the codebase for test functions",
-            },
-            contextItems: [],
-          },
-        ],
-        hasReasoningEnabled: false,
-        isStreaming: true,
-        id: "session-123",
-        mode: "chat",
-        streamAborter: new AbortController(),
-        contextPercentage: 0,
-        isPruned: false,
-        isInEdit: false,
-        title: "",
-        lastSessionId: undefined,
-        isSessionMetadataLoading: false,
-        allSessionMetadata: [],
-        symbols: {},
-        codeBlockApplyStates: {
-          states: [],
-          curIndex: 0,
-        },
-        newestToolbarPreviewForInput: {},
-        compactionLoading: {},
-        inlineErrorMessage: undefined,
-      },
-      config: {
-        config: {
-          tools: [],
-          rules: [],
-          tabAutocompleteModel: undefined,
-          selectedModelByRole: {
-            chat: mockClaudeModel,
-            apply: null,
-            edit: null,
-            summarize: null,
-            autocomplete: null,
-            rerank: null,
-            embed: null,
-          },
-          experimental: {
-            onlyUseSystemMessageTools: false,
-          },
-        } as any,
-        lastSelectedModelByRole: {
-          chat: mockClaudeModel.title,
-        },
-      } as any,
-      ui: {
-        toolSettings: {
-          search_codebase: "askFirst", // Requires manual approval
-        },
-        ruleSettings: {},
-        showDialog: false,
-        dialogMessage: undefined,
-      } as any,
-    });
-
-    const mockIdeMessengerApproval = mockStoreWithApproval.mockIdeMessenger;
-
-    // Setup successful compilation
-    mockIdeMessengerApproval.request.mockImplementation(
-      (endpoint: string, data: any) => {
-        if (endpoint === "llm/compileChat") {
-          return Promise.resolve({
-            status: "success",
-            content: {
-              compiledChatMessages: [
-                {
-                  role: "user",
-                  content: "Please search the codebase for test functions",
-                },
-              ],
-              didPrune: false,
-              contextPercentage: 0.85,
-            },
-          });
-        } else if (endpoint === "tools/call") {
-          // Mock server-side tool call response for search_codebase
-          return Promise.resolve({
-            status: "success",
-            content: {
-              contextItems: [
-                {
-                  name: "Search Results",
-                  description: "Found test functions",
-                  content:
-                    "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
-                  icon: "search",
-                  hidden: false,
-                },
-              ],
-              errorMessage: undefined,
-            },
-          });
-        } else if (endpoint === "history/save") {
-          return Promise.resolve({ status: "success" });
-        } else if (endpoint === "history/list") {
-          return Promise.resolve({ status: "success", content: [] });
-        }
-        return Promise.resolve({ status: "success", content: {} });
-      },
-    );
-
-    // Setup streaming generator with tool call
-    async function* mockStreamGeneratorWithApprovalFlow() {
-      yield [
-        {
-          role: "assistant",
-          content: "I'll search for test functions in the codebase.",
-        },
-      ];
-      yield [
-        {
-          role: "assistant",
-          content: "",
-          toolCalls: [
-            {
-              id: "tool-approval-flow-1",
-              type: "function",
-              function: {
-                name: "search_codebase",
-                arguments: JSON.stringify({ query: "test function" }),
-              },
-            },
-          ],
-        },
-      ];
-      return {
-        prompt: "Please search the codebase for test functions",
-        completion: "I'll search for test functions in the codebase.",
-        modelProvider: "anthropic",
-      };
-    }
-
-    // Mock subsequent streaming calls (after tool execution)
-    let streamCallCount = 0;
-    mockIdeMessengerApproval.llmStreamChat.mockImplementation(() => {
-      streamCallCount++;
-      if (streamCallCount === 1) {
-        // First call - initial streaming with tool call
-        return mockStreamGeneratorWithApprovalFlow();
-      } else {
-        // Subsequent calls from streamResponseAfterToolCall
-        async function* followupGenerator() {
-          yield [
-            {
-              role: "assistant",
-              content:
-                "I found several test functions in your codebase. Here are the main ones I discovered...",
-            },
-          ];
-          return {
-            prompt: "continuing after tool execution",
-            completion:
-              "I found several test functions in your codebase. Here are the main ones I discovered...",
-            modelProvider: "anthropic",
-          };
-        }
-        return followupGenerator();
-      }
-    });
-
-    // Execute initial thunk - this should generate the tool call but not execute it
-    const initialResult = await mockStoreWithApproval.dispatch(
-      streamResponseThunk({
-        editorState: mockEditorState,
-        modifiers: mockModifiers,
-      }) as any,
-    );
-
-    // Verify initial streaming completed successfully
-    expect(initialResult.type).toBe("chat/streamResponse/fulfilled");
-
-    // Verify exact initial action sequence by comparing action types
-    const initialActions = (mockStoreWithApproval as any).getActions();
-
-    expect(initialActions).toEqual([
-      {
-        type: "chat/streamResponse/pending",
-        meta: {
-          arg: {
-            editorState: mockEditorState,
-            modifiers: mockModifiers,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/pending",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/submitEditorAndInitAtIndex",
-        payload: {
-          editorState: mockEditorState,
-          index: 1,
-        },
-      },
-      {
-        type: "session/resetNextCodeBlockToApplyIndex",
-        payload: undefined,
-      },
-      {
-        type: "symbols/updateFromContextItems/pending",
-        meta: {
-          arg: [],
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/updateHistoryItemAtIndex",
-        payload: {
-          index: 1,
-          updates: {
-            contextItems: [],
-            message: {
-              content: "Hello, please help me with this code",
-              id: "mock-uuid-123",
-              role: "user",
-            },
-          },
-        },
-      },
-      {
-        type: "chat/streamNormalInput/pending",
-        meta: {
-          arg: {
-            legacySlashCommandData: undefined,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setAppliedRulesAtIndex",
-        payload: {
-          appliedRules: [],
-          index: 1,
-        },
-      },
-      {
-        type: "session/setActive",
-        payload: undefined,
-      },
-      {
-        type: "session/setInlineErrorMessage",
-        payload: undefined,
-      },
-      {
-        type: "session/setIsPruned",
-        payload: false,
-      },
-      {
-        type: "session/setContextPercentage",
-        payload: 0.85,
-      },
-      {
-        type: "symbols/updateFromContextItems/fulfilled",
-        meta: {
-          arg: [],
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content: "I'll search for test functions in the codebase.",
-          },
-        ],
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content: "",
-            toolCalls: [
-              {
-                id: "tool-approval-flow-1",
-                type: "function",
-                function: {
-                  name: "search_codebase",
-                  arguments: JSON.stringify({ query: "test function" }),
-                },
-              },
-            ],
-          },
-        ],
-      },
-      {
-        type: "session/addPromptCompletionPair",
-        payload: [
-          {
-            completion: "I'll search for test functions in the codebase.",
-            modelProvider: "anthropic",
-            prompt: "Please search the codebase for test functions",
-          },
-        ],
-      },
-      {
-        type: "session/setInactive",
-        payload: undefined,
-      },
-      {
-        type: "session/setToolGenerated",
-        payload: {
-          toolCallId: "tool-approval-flow-1",
-          tools: [],
-        },
-      },
-      {
-        type: "chat/streamNormalInput/fulfilled",
-        meta: {
-          arg: {
-            legacySlashCommandData: undefined,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/pending",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/update/pending",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/updateSessionMetadata",
-        payload: {
-          sessionId: "session-123",
-          title: "New Session",
-        },
-      },
-      {
-        type: "session/refreshMetadata/pending",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setIsSessionMetadataLoading",
-        payload: false,
-      },
-      {
-        type: "session/setAllSessionMetadata",
-        payload: {},
-      },
-      {
-        type: "session/refreshMetadata/fulfilled",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: {},
-      },
-      {
-        type: "session/update/fulfilled",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/fulfilled",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setAllSessionMetadata",
-        payload: [],
-      },
-      {
-        type: "session/refreshMetadata/fulfilled",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: [],
-      },
-      {
-        type: "session/update/fulfilled",
-        meta: {
-          arg: expect.objectContaining({
-            history: expect.any(Array),
-            sessionId: "session-123",
-            title: "New Session",
-            workspaceDirectory: "",
-          }),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/saveCurrent/fulfilled",
-        meta: {
-          arg: {
-            generateTitle: true,
-            openNewSession: false,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/fulfilled",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamResponse/fulfilled",
-        meta: {
-          arg: {
-            editorState: mockEditorState,
-            modifiers: mockModifiers,
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-    ]);
-
-    // Clear the actions array to track only the approval flow
-    (mockStoreWithApproval as any).clearActions();
-
-    // Import the callToolById thunk to simulate user approval
-    const { callToolById } = await import("./callToolById");
-
-    // Simulate user clicking "Accept" on the tool call
-    const approvalResult = await mockStoreWithApproval.dispatch(
-      callToolById({ toolCallId: "tool-approval-flow-1" }) as any,
-    );
-
-    // Verify tool execution completed successfully
-    expect(approvalResult.type).toBe("chat/callTool/fulfilled");
-
-    // Verify exact approval flow actions
-    const approvalActions = (mockStoreWithApproval as any).getActions();
-    expect(approvalActions).toEqual([
-      {
-        type: "chat/callTool/pending",
-        meta: {
-          arg: { toolCallId: "tool-approval-flow-1" },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setToolCallCalling",
-        payload: { toolCallId: "tool-approval-flow-1" },
-      },
-      {
-        type: "session/updateToolCallOutput",
-        payload: {
-          toolCallId: "tool-approval-flow-1",
-          contextItems: [
-            {
-              name: "Search Results",
-              description: "Found test functions",
-              content:
-                "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
-              icon: "search",
-              hidden: false,
-            },
-          ],
-        },
-      },
-      {
-        type: "session/acceptToolCall",
-        payload: { toolCallId: "tool-approval-flow-1" },
-      },
-      {
-        type: "chat/streamAfterToolCall/pending",
-        meta: {
-          arg: {
-            toolCallId: "tool-approval-flow-1",
-          },
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/pending",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/resetNextCodeBlockToApplyIndex",
-        payload: undefined,
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            content:
-              "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
-            role: "tool",
-            toolCallId: "tool-approval-flow-1",
-          },
-        ],
-      },
-      {
-        type: "chat/streamNormalInput/pending",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "pending",
-        },
-        payload: undefined,
-      },
-      {
-        type: "session/setAppliedRulesAtIndex",
-        payload: {
-          appliedRules: [],
-          index: 1,
-        },
-      },
-      {
-        type: "session/setActive",
-        payload: undefined,
-      },
-      {
-        type: "session/setInlineErrorMessage",
-        payload: undefined,
-      },
-      {
-        type: "session/setIsPruned",
-        payload: false,
-      },
-      {
-        type: "session/setContextPercentage",
-        payload: 0.85,
-      },
-      {
-        type: "session/streamUpdate",
-        payload: [
-          {
-            role: "assistant",
-            content:
-              "I found several test functions in your codebase. Here are the main ones I discovered...",
-          },
-        ],
-      },
-      {
-        type: "session/addPromptCompletionPair",
-        payload: [
-          {
-            completion:
-              "I found several test functions in your codebase. Here are the main ones I discovered...",
-            modelProvider: "anthropic",
-            prompt: "continuing after tool execution",
-          },
-        ],
-      },
-      {
-        type: "session/setInactive",
-        payload: undefined,
-      },
-      {
-        type: "chat/streamNormalInput/fulfilled",
-        meta: {
-          arg: {},
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamWrapper/fulfilled",
-        meta: {
-          arg: expect.any(Function),
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/streamAfterToolCall/fulfilled",
-        meta: {
-          arg: {
-            toolCallId: "tool-approval-flow-1",
-          },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-      {
-        type: "chat/callTool/fulfilled",
-        meta: {
-          arg: { toolCallId: "tool-approval-flow-1" },
-          requestId: expect.any(String),
-          requestStatus: "fulfilled",
-        },
-        payload: undefined,
-      },
-    ]);
-
-    // Verify IDE messenger calls for tool execution
-    expect(mockIdeMessengerApproval.request).toHaveBeenCalledWith(
-      "tools/call",
-      {
-        toolCall: {
-          id: "tool-approval-flow-1",
-          type: "function",
-          function: {
-            name: "search_codebase",
-            arguments: JSON.stringify({ query: "test function" }),
-          },
-        },
-      },
-    );
-
-    // Verify second streaming call was made for continuation
-    expect(streamCallCount).toBe(2);
-
-    // Verify final state shows completed tool call and follow-up response
-    const finalState = mockStoreWithApproval.getState();
-    expect(finalState).toEqual({
-      session: {
-        history: [
-          {
-            contextItems: [],
-            message: {
-              id: "1",
-              role: "user",
-              content: "Please search the codebase for test functions",
-            },
-          },
-          {
-            appliedRules: [],
-            contextItems: [],
-            editorState: mockEditorState,
-            message: {
-              content: "Hello, please help me with this code",
-              id: expect.any(String),
-              role: "user",
-            },
-          },
-          {
-            contextItems: [],
-            isGatheringContext: false,
-            message: {
-              content: "I'll search for test functions in the codebase.",
-              id: expect.any(String),
-              role: "assistant",
-              toolCalls: [
-                {
-                  id: "tool-approval-flow-1",
-                  type: "function",
-                  function: {
-                    name: "search_codebase",
-                    arguments: JSON.stringify({ query: "test function" }),
-                  },
-                },
-              ],
-            },
-            promptLogs: [
-              {
-                completion: "I'll search for test functions in the codebase.",
-                modelProvider: "anthropic",
-                prompt: "Please search the codebase for test functions",
-              },
-            ],
-            toolCallStates: [
-              {
-                toolCallId: "tool-approval-flow-1",
-                toolCall: {
-                  id: "tool-approval-flow-1",
-                  type: "function",
-                  function: {
-                    name: "search_codebase",
-                    arguments: JSON.stringify({ query: "test function" }),
-                  },
-                },
-                parsedArgs: { query: "test function" },
-                status: "done", // Tool call completed successfully
-                output: [
-                  {
-                    name: "Search Results",
-                    description: "Found test functions",
-                    content:
-                      "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
-                    icon: "search",
-                    hidden: false,
-                  },
-                ],
-              },
-            ],
-          },
-          {
-            contextItems: [],
-            message: {
-              content:
-                "function testUserLogin() {...}\\nfunction testDataValidation() {...}",
-              id: expect.any(String),
-              role: "tool",
-              toolCallId: "tool-approval-flow-1",
-            },
-          },
-          {
-            contextItems: [],
-            isGatheringContext: false,
-            message: {
-              content:
-                "I found several test functions in your codebase. Here are the main ones I discovered...",
-              id: "mock-uuid-123",
-              role: "assistant",
-            },
-            promptLogs: [
-              {
-                completion:
-                  "I found several test functions in your codebase. Here are the main ones I discovered...",
-                modelProvider: "anthropic",
-                prompt: "continuing after tool execution",
-              },
-            ],
-          },
-        ],
-        hasReasoningEnabled: false,
-        isStreaming: false, // Inactive after complete flow
-        id: "session-123",
-        mode: "chat",
-        streamAborter: expect.any(AbortController),
-        contextPercentage: 0.85,
-        isPruned: false,
-        isInEdit: false,
-        title: "New Session",
-        lastSessionId: undefined,
-        isSessionMetadataLoading: false,
-        allSessionMetadata: {},
-        symbols: {},
-        codeBlockApplyStates: {
-          states: [],
-          curIndex: 0,
-        },
-        newestToolbarPreviewForInput: {},
-        compactionLoading: {},
-        inlineErrorMessage: undefined,
-      },
-      config: {
-        config: {
-          tools: [],
-          rules: [],
-          tabAutocompleteModel: undefined,
-          selectedModelByRole: {
-            chat: mockClaudeModel,
-            apply: null,
-            edit: null,
-            summarize: null,
-            autocomplete: null,
-            rerank: null,
-            embed: null,
-          },
-          experimental: {
-            onlyUseSystemMessageTools: false,
-          },
-        },
-        lastSelectedModelByRole: {
-          chat: mockClaudeModel.title,
-        },
-        loading: false,
-        configError: undefined,
-      },
-      ui: {
-        toolSettings: {
-          search_codebase: "askFirst",
-        },
-        ruleSettings: {},
-        showDialog: false,
-        dialogMessage: undefined,
-      },
-      editModeState: {
-        isInEdit: false,
-        returnToMode: "chat",
-      },
-      indexing: {
-        indexingState: "disabled",
-      },
-      tabs: {
-        tabsItems: [],
-      },
-      profiles: {
-        profiles: [],
-      },
-    });
-  });
 });
